// solar-frontend/src/app/+dealers/models/dealer.ts
/**
 * This is Dealer interface to give data type for Dealer object
 */

export interface IDealer {
  dealerCd: string;
  city: string;
  affil: string;
  shipId: string;
  desc: string;
  zipCode: string;
  phone: string;
  phone2: string;
  state: string;
  email1: string;
  email2: string;
  contact: string;
  longtitude: number;
  latitude: number;
  primaryEmailFlag: string;
  primaryPhoneFlag: string;
  primaryEmailFlag2: string;
  primaryPhoneFlag2: string;
  instruction: string;
  instruction2: string;
  address: string;
}